# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateCountsResponse": {
        "properties": {
          "dense": {
            "title": "Dense",
            "type": "integer"
          },
          "final": {
            "title": "Final",
            "type": "integer"
          },
          "fused": {
            "title": "Fused",
            "type": "integer"
          },
          "keyword": {
            "title": "Keyword",
            "type": "integer"
          },
          "reranked": {
            "title": "Reranked",
            "type": "integer"
          }
        },
        "required": [
          "keyword",
          "dense",
          "fused",
          "reranked",
          "final"
        ],
        "title": "CandidateCountsResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Standard error response schema published in OpenAPI.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RetrievalMode": {
        "enum": [
          "keyword",
          "dense",
          "dense_rerank",
          "hybrid",
          "hybrid_rerank"
        ],
        "title": "RetrievalMode",
        "type": "string"
      },
...
      "SearchApiRequest": {
        "additionalProperties": false,
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/SearchFiltersRequest"
          },
          "mode": {
            "$ref": "#/components/schemas/RetrievalMode",
            "default": "dense_rerank"
          },
          "query": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchApiRequest",
        "type": "object"
      },
      "SearchApiResponse": {
        "properties": {
          "degradation_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Degradation Reasons",
            "type": "array"
          },
          "degraded": {
            "title": "Degraded",
            "type": "boolean"
          },
          "dense_index_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dense Index Version"
          },
          "diagnostics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchDiagnosticsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "embedding_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Version"
          },
          "executed_mode": {
            "$ref": "#/components/schemas/RetrievalMode"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/SearchHitResponse"
            },
            "title": "Hits",
            "type": "array"
          },
          "keyword_index_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyword Index Version"
          },
          "profile_version": {
            "title": "Profile Version",
            "type": "string"
          },
          "requested_mode": {
            "$ref": "#/components/schemas/RetrievalMode"
          },
          "reranker_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reranker Version"
          }
        },
        "required": [
          "requested_mode",
          "executed_mode",
          "profile_version",
          "embedding_version",
          "reranker_version",
          "keyword_index_version",
          "dense_index_version",
          "degraded",
          "degradation_reasons",
          "hits"
        ],
        "title": "SearchApiResponse",
        "type": "object"
      },
      "SearchDiagnosticsResponse": {
        "properties": {
          "candidate_counts": {
            "$ref": "#/components/schemas/CandidateCountsResponse"
          },
          "context_only_count": {
            "title": "Context Only Count",
            "type": "integer"
          },
          "degradation_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Degradation Reasons",
            "type": "array"
          },
          "degraded": {
            "title": "Degraded",
            "type": "boolean"
          },
          "dense_index_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dense Index Version"
          },
          "dense_language_slice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dense Language Slice"
          },
          "dense_query_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dense Query Kind"
          },
          "document_filter_count": {
            "title": "Document Filter Count",
            "type": "integer"
          },
          "embedding_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Version"
          },
          "executed_mode": {
            "$ref": "#/components/schemas/RetrievalMode"
          },
          "filter_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Filter Reasons",
            "type": "array"
          },
          "keyword_index_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyword Index Version"
          },
          "keyword_language_slice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyword Language Slice"
          },
          "keyword_literal_term_count": {
            "title": "Keyword Literal Term Count",
            "type": "integer"
          },
          "keyword_query_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyword Query Kind"
          },
          "profile_version": {
            "title": "Profile Version",
            "type": "string"
          },
          "requested_mode": {
            "$ref": "#/components/schemas/RetrievalMode"
          },
          "reranker_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reranker Version"
          },
          "source_filter_count": {
            "title": "Source Filter Count",
            "type": "integer"
          },
          "stage_timings": {
            "items": {
              "$ref": "#/components/schemas/StageTimingResponse"
            },
            "title": "Stage Timings",
            "type": "array"
          }
        },
        "required": [
          "requested_mode",
          "executed_mode",
          "profile_version",
          "embedding_version",
          "reranker_version",
          "keyword_index_version",
          "dense_index_version",
          "candidate_counts",
          "stage_timings",
          "keyword_language_slice",
          "keyword_query_kind",
          "keyword_literal_term_count",
          "dense_language_slice",
          "dense_query_kind",
          "degraded",
          "degradation_reasons",
          "filter_reasons",
          "source_filter_count",
          "document_filter_count",
          "context_only_count"
        ],
        "title": "SearchDiagnosticsResponse",
        "type": "object"
      },
      "SearchFiltersRequest": {
        "additionalProperties": false,
        "properties": {
          "document_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 100,
            "title": "Document Ids",
            "type": "array"
          },
          "source_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 100,
            "title": "Source Ids",
            "type": "array"
          }
        },
        "title": "SearchFiltersRequest",
        "type": "object"
      },
...
      "SpaceCreateRequest": {
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "SpaceCreateRequest",
        "type": "object"
      },
      "SpaceResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_id": {
            "title": "Owner Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "owner_id"
        ],
        "title": "SpaceResponse",
        "type": "object"
      },
      "StageTimingResponse": {
        "properties": {
          "latency_ms": {
            "title": "Latency Ms",
            "type": "number"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "stage",
          "latency_ms"
        ],
        "title": "StageTimingResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Agent Knowledge Repository",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/spaces": {
      "get": {
        "operationId": "list_spaces_api_v1_spaces_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SpaceResponse"
                  },
                  "title": "Response List Spaces Api V1 Spaces Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Spaces",
        "tags": [
          "spaces"
        ]
      },
      "post": {
        "operationId": "create_space_api_v1_spaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpaceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Space",
        "tags": [
          "spaces"
        ]
      }
    },
    "/api/v1/spaces/{space_id}": {
      "delete": {
        "operationId": "delete_space_api_v1_spaces__space_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Space",
        "tags": [
          "spaces"
        ]
      }
    },
...
    "/api/v1/spaces/{space_id}/search": {
      "post": {
        "description": "Execute one bounded retrieval request inside a single Space.",
        "operationId": "search_space_api_v1_spaces__space_id__search_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchApiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Stable retrieval error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Stable retrieval error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Stable retrieval error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Stable retrieval error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Stable retrieval error"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Stable retrieval error"
          }
        },
        "summary": "Search Space"
      }
    },
...
  }
}